+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - When holding a Redstone Tool, linked positions show as blue instead of red now
   - That is, the block the Tool is linked to is still red. Blocks that the linked blook is linked to are blue
 
+### Fixed
+
+- The Redstone Tool would show an indicator on the linked block twice
+
 # 1.4.0
 
 ### Changed